+++ client/src/elements/schema-editor/schema-editor-boolean.js
@@ -1,7 +1,12 @@
 import { bindable } from "aurelia-framework";
 
 export class SchemaEditorBoolean {
-  @bindable schema;
-  @bindable data;
-  @bindable change;
+  @bindable
+  schema;
+  @bindable
+  data;
+  @bindable
+  change;
+  @bindable
+  showNotifications;
 }