+++ client/src/elements/schema-editor/schema-editor-json.js
@@ -4,9 +4,18 @@ import Ajv from "ajv";
 const ajv = new Ajv();
 
 export class SchemaEditorJson {
-  @bindable schema;
-  @bindable data;
-  @bindable change;
+  @bindable
+  schema;
+  @bindable
+  data;
+  @bindable
+  change;
+  @bindable
+  required;
+  @bindable
+  showNotifications;
+
+  options = {};
 
   constructor() {
     this.handleChange = () => {
@@ -20,6 +29,13 @@ export class SchemaEditorJson {
   schemaChanged() {
     if (this.schema) {
       this.ajvValidate = ajv.compile(this.schema);
+      this.applyOptions();
+    }
+  }
+
+  applyOptions() {
+    if (this.schema.hasOwnProperty("Q:options")) {
+      this.options = Object.assign(this.options, this.schema["Q:options"]);
     }
   }