+++ client/src/resources/availability-checks/ToolEndpointAvailabilityCheck.js
@@ -2,30 +2,21 @@ import { inject, LogManager } from "aurelia-framework";
 import ToolEndpointChecker from "resources/checkers/ToolEndpointChecker.js";
 
 const log = LogManager.getLogger("Q");
-
 @inject(ToolEndpointChecker)
 export default class ToolEndpointAvailabilityCheck {
   constructor(toolEndpointChecker) {
     this.toolEndpointChecker = toolEndpointChecker;
   }
 
-  async isAvailable(availabilityCheck) {
-    if (!availabilityCheck.endpoint) {
+  async isAvailable(config) {
+    if (!config.endpoint) {
       log.error(
         "no endpoint defined for availabilityCheck checkToolEndpoint:",
-        availabilityCheck
+        config
       );
       return false;
     }
-    if (!availabilityCheck.withData) {
-      const availability = await this.toolEndpointChecker.fetch(
-        availabilityCheck.endpoint
-      );
-      return availability.available;
-    }
-    const availability = await this.toolEndpointChecker.fetchWithItem(
-      availabilityCheck.endpoint
-    );
-    return availability.available;
+    const result = await this.toolEndpointChecker.check(config);
+    return result.available;
   }
 }