+++ client/src/resources/checkers/AvailabilityChecker.js
@@ -1,5 +1,22 @@
-import { inject } from "aurelia-framework";
+import { inject, LogManager } from "aurelia-framework";
 import { Container } from "aurelia-dependency-injection";
+
+const log = LogManager.getLogger("Q");
+
+// This function transforms the existing check config to the new format
+// After all the tools adopted the new configuration format this is
+// not needed anymore
+function getConfig(availabilityCheck) {
+  const check = JSON.parse(JSON.stringify(availabilityCheck));
+  if (check.config) {
+    return check.config;
+  }
+  delete check.type;
+  log.info(
+    "DEPRECATION NOTICE: In Q editor 4.0 you will have to configure the availabilityCheck with a config property. See https://github.com/nzzdev/Q-editor/blob/master/README.md for details"
+  );
+  return check;
+}
 
 @inject(Container)
 export default class AvailabilityChecker {
@@ -33,7 +50,10 @@ export default class AvailabilityChecker {
         let checker = this.diContainer.get(
           availabilityCheck.type + "AvailabilityCheck"
         );
-        const available = await checker.isAvailable(availabilityCheck);
+
+        const available = await checker.isAvailable(
+          getConfig(availabilityCheck)
+        );
         if (!available) {
           const availabilityInfo = {
             isAvailable: false