+++ client/src/resources/notification-checks/ToolEndpointNotificationCheck.js
@@ -9,28 +9,14 @@ export default class ToolEndpointNotificationCheck {
     this.toolEndpointChecker = toolEndpointChecker;
   }
 
-  async getNotification(notificationCheck, data) {
-    if (!notificationCheck.endpoint) {
+  async getNotification(config) {
+    if (!config.endpoint) {
       log.error(
         "no endpoint defined for notificationCheck ToolEndpointNotificationCheck:",
-        notificationCheck
+        config
       );
       return false;
     }
-    const dataForEndpoint = {
-      data: data
-    };
-    if (notificationCheck.hasOwnProperty("options")) {
-      dataForEndpoint.options = notificationCheck.options;
-    }
-    try {
-      const notification = await this.toolEndpointChecker.fetchWithData(
-        notificationCheck.endpoint,
-        dataForEndpoint
-      );
-      return notification;
-    } catch (e) {
-      log.error("failed to fetch with toolEndpointChecker", e);
-    }
+    return await this.toolEndpointChecker.check(config);
   }
 }