+++ client/src/resources/notification-checks/index.js
@@ -2,7 +2,6 @@ import EmptyDataNotificationCheck from "./EmptyDataNotificationCheck";
 import HasColumnTitlesNotificationCheck from "./HasColumnTitlesNotificationCheck.js";
 import TooManyColumnsNotificationCheck from "./TooManyColumnsNotificationCheck.js";
 import ToolEndpointNotificationCheck from "./ToolEndpointNotificationCheck.js";
-import IsValueMissingNotificationCheck from "./IsValueMissingNotificationCheck.js";
 
 export function configure(frameworkConfiguration) {
   frameworkConfiguration.singleton(
@@ -20,9 +19,5 @@ export function configure(frameworkConfiguration) {
   frameworkConfiguration.singleton(
     "ToolEndpointNotificationCheck",
     ToolEndpointNotificationCheck
-  );
-  frameworkConfiguration.singleton(
-    "IsValueMissingNotificationCheck",
-    IsValueMissingNotificationCheck
   );
 }