+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+# 3.6.0 (14.11.2018)
+- feature: required notifications in editor are shown using the notification system instead of browser native notification
+- feature: availability-checks, notification-checks and dynamic-enum config format is unified (see README for examples)
+- fix: schema-editor-files handles thumbnail correctly if no file type is given
+- refactor: livingdocs-component code is now much cleaner
+
 # 3.5.3 (24.10.2018)
 - fix: trim all cells in schema-editor-table when data is pasted or edited by hand